split utility functions and memory manager into separate files
