+++ system/camerad/cameras/camera_common.cc
@@ -102,16 +102,3 @@ float calculate_exposure_value(const CameraBuf *b, Rect ae_xywh, int x_skip, int
 
   return lum_med / 256.0;
 }
-
-int open_v4l_by_name_and_index(const char name[], int index, int flags) {
-  for (int v4l_index = 0; /**/; ++v4l_index) {
-    std::string v4l_name = util::read_file(util::string_format("/sys/class/video4linux/v4l-subdev%d/name", v4l_index));
-    if (v4l_name.empty()) return -1;
-    if (v4l_name.find(name) == 0) {
-      if (index == 0) {
-        return HANDLE_EINTR(open(util::string_format("/dev/v4l-subdev%d", v4l_index).c_str(), flags));
-      }
-      index--;
-    }
-  }
-}
