+++ system/camerad/cameras/spectra.h
@@ -1,10 +1,6 @@
 #pragma once
 
-#include <sys/mman.h>
-#include <functional>
 #include <memory>
-#include <queue>
-#include <optional>
 #include <utility>
 
 #include "media/cam_req_mgr.h"
@@ -12,6 +8,7 @@
 #include "common/util.h"
 #include "system/camerad/cameras/hw.h"
 #include "system/camerad/cameras/camera_common.h"
+#include "system/camerad/cameras/mem_mgr.h"
 #include "system/camerad/sensors/sensor.h"
 
 #define MAX_IFE_BUFS 20
@@ -36,34 +33,6 @@ typedef enum {
   ISP_BPS_PROCESSED,  // fully processed image through the BPS
 } SpectraOutputType;
 
-std::optional<int32_t> device_acquire(int fd, int32_t session_handle, void *data, uint32_t num_resources=1);
-int device_config(int fd, int32_t session_handle, int32_t dev_handle, uint64_t packet_handle);
-int device_control(int fd, int op_code, int session_handle, int dev_handle);
-int do_cam_control(int fd, int op_code, void *handle, int size);
-void *alloc_w_mmu_hdl(int video0_fd, int len, uint32_t *handle, int align = 8, int flags = CAM_MEM_FLAG_KMD_ACCESS | CAM_MEM_FLAG_UMD_ACCESS | CAM_MEM_FLAG_CMD_BUF_TYPE,
-                      int mmu_hdl = 0, int mmu_hdl2 = 0);
-void release(int video0_fd, uint32_t handle);
-
-class MemoryManager {
-public:
-  void init(int _video0_fd) { video0_fd = _video0_fd; }
-  ~MemoryManager();
-
-  template <class T>
-  auto alloc(int len, uint32_t *handle) {
-    return std::unique_ptr<T, std::function<void(void *)>>((T*)alloc_buf(len, handle), [this](void *ptr) { this->free(ptr); });
-  }
-
-private:
-  void *alloc_buf(int len, uint32_t *handle);
-  void free(void *ptr);
-
-  std::map<void *, uint32_t> handle_lookup;
-  std::map<void *, int> size_lookup;
-  std::map<int, std::queue<void *> > cached_allocations;
-  int video0_fd;
-};
-
 class SpectraMaster {
 public:
   void init();
@@ -76,42 +45,6 @@ public:
   int cdm_iommu = -1;
   int icp_device_iommu = -1;
   MemoryManager mem_mgr;
-};
-
-class SpectraBuf {
-public:
-  SpectraBuf() = default;
-
-  ~SpectraBuf() {
-    if (video_fd >= 0 && ptr) {
-      munmap(ptr, mmap_size);
-      release(video_fd, handle);
-    }
-  }
-
-  void init(SpectraMaster *m, int s, int a, bool shared_access, int mmu_hdl = 0, int mmu_hdl2 = 0, int count = 1) {
-    video_fd = m->video0_fd;
-    size = s;
-    alignment = a;
-    mmap_size = aligned_size() * count;
-
-    uint32_t flags = CAM_MEM_FLAG_HW_READ_WRITE | CAM_MEM_FLAG_KMD_ACCESS | CAM_MEM_FLAG_UMD_ACCESS | CAM_MEM_FLAG_CMD_BUF_TYPE;
-    if (shared_access) {
-      flags |= CAM_MEM_FLAG_HW_SHARED_ACCESS;
-    }
-
-    void *p = alloc_w_mmu_hdl(video_fd, mmap_size, (uint32_t*)&handle, alignment, flags, mmu_hdl, mmu_hdl2);
-    ptr = (unsigned char*)p;
-    assert(ptr != NULL);
-  };
-
-  uint32_t aligned_size() {
-    return ALIGNED_SIZE(size, alignment);
-  };
-
-  int video_fd = -1;
-  unsigned char *ptr = nullptr;
-  int size = 0, alignment = 0, handle = 0, mmap_size = 0;
 };
 
 class SpectraCamera {
